requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pandas
numpy
xgboost
scikit-learn
openpyxl
opencv-python
pytesseract
spacy
motor
boto3
python-dotenv
python-jose
passlib[bcrypt]
redis
arq
